+++ Source/aweXpect.Core/Customization/CustomizationLifetime.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace aweXpect.Customization;
+
+/// <summary>
+///     The lifetime of a customization setting.
+/// </summary>
+public sealed class CustomizationLifetime(Action callback) : IDisposable
+{
+	/// <inheritdoc cref="IDisposable.Dispose()" />
+	public void Dispose() => callback();
+}
